Issue #2818001 by stevector: Refactor for SolrConnector Plugin
