+++ tests/behat/features/server.feature
@@ -8,7 +8,8 @@ Feature: Solr on Pantheon
     When I visit "admin/config/search/search-api/add-server"
     And I fill in "name" with "pantheon"
     And I fill in "id" with "pantheon"
-    And I select the radio button "Solr on Pantheon"
+    And I select the radio button "Solr"
+    And I select the radio button "Pantheon"
     And I press the "Save" button
     And I select the radio button "modules/search_api_solr/solr-conf/4.x/schema.xml"
     And I press the "Save" button
